move legacy ListItem
