+++ src/screens/LiquidityPools/LiquidityPools.js
@@ -31,7 +31,7 @@ import Image from 'components/Image';
 import { Spacing, ScrollWrapper } from 'components/legacy/Layout';
 import { BaseText } from 'components/legacy/Typography';
 import Tabs from 'components/Tabs';
-import ListItemWithImage from 'components/ListItem/ListItemWithImage';
+import ListItemWithImage from 'components/legacy/ListItem/ListItemWithImage';
 import RefreshControl from 'components/RefreshControl';
 import RetryGraphQueryBox from 'components/RetryGraphQueryBox';
